// apps/local-server/src/integrations/linear/types.ts
export interface LinearIssueSummary {
  id: string;
  ref: string;
  title: string;
  url: string;
}
// ...
export interface LinearResolvedIssue extends LinearIssueSummary {
  blocks: LinearIssueSummary[];
  description: string | null;
  priority: number | null;
  state: {
    name: string;
    type: string;
  } | null;
  project: {
    name: string;
  } | null;
  team: {
    key: string;
    name: string;
  } | null;
}